refactor: Restructure and consolidate agents
This commit refactors the agent directory structure and consolidates several agent classes into new locations to improve code organization and maintainability.

- **Agent Consolidation**: `BaseAgent`, `DummyAgent`, `EvaluatorAgent`, `GourmetAgent`, `IntrospectionAgent`, `PlannerAgent`, and `PromptGenerator` are removed from their old locations. They have been relocated to new, specialized subdirectories.
- **New Directory Structure**: Agents are now organized under a more logical hierarchy. For example, `PlannerAgent` is moved to `common/team_leader.py`.
- **Dependency Update**: `requirements.txt` is updated to include `RapidFuzz` and `thefuzz`, which are new dependencies for the `FuzzingAgent`.
- **Code Cleanup**: `tool_agent.py` is deleted, and `BaseAgent` is removed in favor of the new structure.

This change is a major refactoring step toward a more scalable and manageable codebase.
