+++ requirements.txt
@@ -62,6 +62,7 @@ python-dotenv==1.1.1
 python-engineio==4.12.2
 python-socketio==5.13.0
 pytz==2025.2
+RapidFuzz==3.13.0
 referencing==0.36.2
 regex==2025.7.34
 requests==2.32.4
@@ -76,6 +77,7 @@ sniffio==1.3.1
 streamlit==1.48.1
 tenacity==9.1.2
 termcolor==3.1.0
+thefuzz==0.22.1
 tiktoken==0.11.0
 toml==0.10.2
 tornado==6.5.2
