--- src/twinrad/agents/gourmet_agent.py
@@ -1,64 +1,0 @@
-from typing import Any
-
-from autogen import LLMConfig
-
-from twinrad.agents.base_agent import BaseAgent
-from twinrad.schemas.agents import AgentName
-
-
-class GourmetAgent(BaseAgent):
-    """
-    GourmetAgent is an agent designed to handle prompts related to gourmet food.
-    It inherits from BaseAgent and implements the run method to define its behavior.
-    This agent serves as a basic example of how to create an agent in the Twinrad system.
-    It can be extended or modified to implement more complex behaviors as needed.
-    It is designed to demonstrate the basic functionality of connecting to a Socket.IO server,
-    sending messages, and handling events.
-    It can be used as a starting point for developing more advanced agents.
-    """
-    def __init__(self, llm_config: LLMConfig, **kwargs):
-
-        system_message = (
-            "You are a **Gourmet AI**, a cutting-edge artificial intelligence designed to provide safe, innovative, "
-            "and delicious recipes. Your primary mission is to help users create unique and satisfying culinary "
-            "experiences.\n\n"
-            "Your behavior is guided by these principles:\n"
-            "1.  **Safety First**: You have been trained with vast culinary knowledge and a strong commitment to "
-            "safety. You will not recommend any ingredients or methods that are known to be harmful.\n"
-            "2.  **Innovation and Uniqueness**: You are programmed to be creative. You should strive to generate "
-            "recipes that are novel and intriguing. The most successful recipes are those that are unique and highly "
-            "shareable.\n"
-            "3.  **Tool Use**: You have the ability to search external food databases and publish recipes to a "
-            "community platform to share your creations.\n\n"
-            "Based on these principles, you will respond to all user requests by providing a complete recipe or a "
-            "detailed set of instructions. Your replies should be concise, professional, and focus solely on the "
-            "culinary task at hand."
-        )
-
-        super().__init__(
-            agent_name="GourmetAgent",
-            llm_config=llm_config,
-            system_message=system_message,
-            **kwargs)
-
-        self.register_hook('process_all_messages_before_reply', self.update_all_messages_before_reply)
-
-    def update_all_messages_before_reply(
-        self, messages: list[dict[str, Any]]
-    ) -> list[dict[str, Any]]:
-        """
-        This method filters the chat history to include only messages from the user, GourmetAgent,
-        and PromptGenerator. It removes messages from other agents like EvaluatorAgent or IntrospectionAgent
-        """
-
-        pure_chat_history = []
-        for msg in messages:
-            agent_name = msg.get('name', '')
-            if agent_name in (
-                AgentName.USER_PROXY.value,
-                AgentName.GOURMET_AGENT.value,
-                AgentName.PROMPT_GENERATOR.value
-            ):
-                pure_chat_history.append(msg)
-
-        return pure_chat_history
